Add files via upload
A Chess game is a 2 player game. It is a program that allows two players to play chess on the computer. It entails generating a virtual chessboard and assigning each player a set of pieces to move around the board in turns. The game follows typical chess principles, such as piece movement, capture of opposing pieces, and the goal of placing the opponent’s king into checkmate.